Use common multi module logic to determine which module contains Lib server dir
